classifier/requirements.txt
pandas
tweet-preprocessor
transformers
datasets
peft
torch
scikit-learn
sentence-transformers